// packages/api-client/src/generated-types/ChaptersQueryParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs).
// Relative imports converted to absolute imports by generate-types recipe. DO NOT EDIT.

export type ChaptersQueryParams = {
  search?: string;
  sort_by?: string;
  sort_order?: string;
  page?: bigint;
  limit?: bigint;
};